requirements.txt
# Core dependencies
litellm>=1.30.0
pydantic>=2.5.0
click>=8.1.7
aiohttp>=3.9.0
pyyaml>=6.0.1

# Vector database for code indexing
chromadb>=0.4.22

# Embeddings for semantic search - используем litellm embedding API

# Async utilities
asyncio-mqtt>=0.16.1

# Terminal command execution
pexpect>=4.9.0

# Web search (optional providers)
duckduckgo-search>=4.1.0
tavily-python>=0.3.0

# Development tools
black>=23.12.0
pytest>=7.4.4
pytest-asyncio>=0.21.1

